- Initial work on interfaces - Rework unions a bit to fix support types that aren't specified through a InlineFragment

+++ Sources/SwiftGraphQLGenerator/Primitives.swift
@@ -13,7 +13,7 @@ public enum Primitives: String {
         switch self {
             case .ID: return .string
             case .Int: return .int
-            case .Float: return .float
+            case .Float: return .double
             case .String: return .string
             case .Boolean: return .bool
         }
